# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_readygeneration_readygeneration-backend_pkg_response.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  github_com_readygeneration_readygeneration-backend_pkg_response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.APIError'
      meta:
        $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Meta'
      success:
        type: boolean
    type: object
  github_com_readygeneration_readygeneration-backend_pkg_response.Meta:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  internal_handler.changePasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  internal_handler.createApplicationRequest:
    properties:
      assigned_to:
        type: string
      grant_id:
        type: string
      internal_deadline:
        type: string
      notes:
        type: string
      priority:
        type: string
    required:
    - grant_id
    type: object
...
  internal_handler.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_handler.resetPasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
  internal_handler.signupRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  internal_handler.updateApplicationRequest:
    properties:
      award_amount:
        type: integer
      internal_deadline:
        type: string
      notes:
        type: string
      priority:
        type: string
      stage:
        type: string
      status:
        type: string
      submission_date:
        type: string
    type: object
  internal_handler.updateApplicationStatusRequest:
    properties:
      stage:
        type: string
      status:
        type: string
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Grant intelligence platform API
  title: ReadyGeneration API
  version: "1.0"
paths:
...
  /applications/{application_id}/narratives:
    get:
      parameters:
      - description: Application UUID
        in: path
        name: application_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: List AI narratives for an application
      tags:
      - narratives
  /applications/{id}:
    get:
      parameters:
      - description: Application UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Get a grant application by ID
      tags:
      - applications
    patch:
      consumes:
      - application/json
      parameters:
      - description: Application UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.updateApplicationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Update application fields
      tags:
      - applications
  /applications/{id}/activities:
    get:
      parameters:
      - description: Application UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Get activity log for an application
      tags:
      - applications
  /applications/{id}/narratives:
    get:
      parameters:
      - description: Application UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Get all AI narratives for an application
      tags:
      - applications
  /applications/{id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Application UUID
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.updateApplicationStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Advance or change an application's pipeline status
      tags:
      - applications
  /auth/change-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Change password payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.changePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Change the authenticated user's password
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      summary: Authenticate and receive a JWT
      tags:
      - auth
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Get the authenticated user
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset password payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.resetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      summary: Reset password using a reset token
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: Signup payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.signupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      summary: Register a new user
      tags:
      - auth
...
  /orgs/{org_id}/applications:
    get:
      parameters:
      - description: Organization UUID
        in: path
        name: org_id
        required: true
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: List applications for an organization
      tags:
      - applications
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization UUID
        in: path
        name: org_id
        required: true
        type: string
      - description: Application data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createApplicationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Start tracking a grant application
      tags:
      - applications
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"